requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.26.0
pydantic==2.7.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.7
typing_extensions==4.12.2
